+++ packages/rocketh-diamond/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @rocketh/diamond
 
+## 0.17.5
+
+### Patch Changes
+
+- Updated dependencies
+  - @rocketh/core@0.17.4
+  - @rocketh/deploy@0.17.4
+  - @rocketh/read-execute@0.17.2
+
 ## 0.17.4
 
 ### Patch Changes