+++ packages/rocketh-doc/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @rocketh/doc
 
+## 0.17.5
+
+### Patch Changes
+
+- @rocketh/node@0.17.5
+
 ## 0.17.4
 
 ### Patch Changes