+++ packages/rocketh-router/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @rocketh/router
 
+## 0.17.6
+
+### Patch Changes
+
+- Updated dependencies
+  - @rocketh/core@0.17.4
+  - @rocketh/deploy@0.17.4
+
 ## 0.17.5
 
 ### Patch Changes