+++ packages/rocketh-web/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @rocketh/web
 
+## 0.17.5
+
+### Patch Changes
+
+- Updated dependencies
+  - @rocketh/core@0.17.4
+  - rocketh@0.17.5
+
 ## 0.17.4
 
 ### Patch Changes